Enable country filtering for Nordigen bank list
Enabling the country filter stops the selector loop from working because
instead of a `.ob-list-institution` it starts with `.ob-country-list`
items which then get replaced on click so banks don't exist on load.

By wrapping the institution loop in an observer it'll automatically
attach the events when a country is selected and it works wonderfully.